Final version of tidying
- Remove unecessary Class
- Tidy class
