+++ SQA_AutomationTest/Admin/NhanVien/TimKiemNV.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 
 namespace SQA_AutomationTest.Admin.NhanVien
 {
-    internal class TimKiemNV : BaseTest
+    internal class TimKiemNV : Tests
     {
         [Test]
         [TestCase("bngoc.hi4103@gmail.com", "17012003")]
